+++ FreeAIr/MCP/ServerProxy/VS/Tools/GetAllWarningErrorsTool.cs
@@ -18,7 +18,7 @@ namespace FreeAIr.MCP.McpServerProxy.VS.Tools
                 VisualStudioMcpServerProxy.VisualStudioProxyName,
                 VisualStudioToolName,
                 "Returns a JSON-formatted list of build warning and build errors of current solution. The result of this function includes: a type (error, warning), a description, file path, and line and column where error (warning) was found.",
-                "{}"
+                NoParameters
                 )
         {
         }